tests/wpt/mozilla/tests/bluetooth/stopNotifications/characteristic-is-removed.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}],
        optionalServices: [generic_access.name]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(heart_rate.name))
    .then(service => service.getCharacteristic(heart_rate_measurement.name))
    .then(characteristic => characteristic.startNotifications())
    .then(characteristic => {
        window.testRunner.setBluetoothMockDataSet(adapter_type.missing_characteristic_heart_rate);
        return promise_rejects_dom(t, 'InvalidStateError', characteristic.stopNotifications());
    });
}, 'Characteristic gets removed. Reject with InvalidStateError.');
</script>
